refactor RebalancingTargetCalculator
