+++ contribs/drt/src/main/java/org/matsim/contrib/drt/optimizer/rebalancing/targetcalculator/LinearRebalancingTargetCalculator.java
@@ -1,8 +1,9 @@
-/* *********************************************************************** *
+/*
+ * *********************************************************************** *
  * project: org.matsim.*
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2018 by the members listed in the COPYING,        *
+ * copyright       : (C) 2020 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -14,13 +15,19 @@
  *   (at your option) any later version.                                   *
  *   See also COPYING, LICENSE and WARRANTY file                           *
  *                                                                         *
- * *********************************************************************** */
+ * *********************************************************************** *
+ */
 
-package org.matsim.contrib.drt.optimizer.rebalancing.mincostflow;
+package org.matsim.contrib.drt.optimizer.rebalancing.targetcalculator;
+
+import java.util.List;
+import java.util.Map;
+import java.util.function.ToIntFunction;
 
 import org.matsim.contrib.drt.analysis.zonal.DrtZone;
 import org.matsim.contrib.drt.optimizer.rebalancing.demandestimator.ZonalDemandEstimator;
-import org.matsim.contrib.drt.optimizer.rebalancing.mincostflow.MinCostFlowRebalancingStrategy.RebalancingTargetCalculator;
+import org.matsim.contrib.drt.optimizer.rebalancing.mincostflow.MinCostFlowRebalancingStrategyParams;
+import org.matsim.contrib.dvrp.fleet.DvrpVehicle;
 
 /**
  * @author michalm
@@ -30,19 +37,22 @@ public class LinearRebalancingTargetCalculator implements RebalancingTargetCalcu
 	private final MinCostFlowRebalancingStrategyParams params;
 
 	public LinearRebalancingTargetCalculator(ZonalDemandEstimator demandEstimator,
-											 MinCostFlowRebalancingStrategyParams params) {
+			MinCostFlowRebalancingStrategyParams params) {
 		this.demandEstimator = demandEstimator;
 		this.params = params;
 	}
 
-	// FIXME targets should be calculated more intelligently
 	@Override
-	public int estimate(DrtZone zone, double time) {
+	public ToIntFunction<DrtZone> calculate(double time, Map<DrtZone, List<DvrpVehicle>> rebalancableVehiclesPerZone) {
 		// XXX this "time+60" (taken from old code) means probably "in the next time bin"
-		int expectedDemand = demandEstimator.getExpectedDemandForTimeBin(time + 60).applyAsInt(zone);
-		if (expectedDemand == 0) {
-			return 0;// for larger zones we may assume that target is at least 1 (or in some cases) ??????
-		}
-		return (int)Math.round(params.getTargetAlpha() * expectedDemand + params.getTargetBeta());
+		ToIntFunction<DrtZone> expectedDemandFunction = demandEstimator.getExpectedDemandForTimeBin(time + 60);
+
+		return zone -> {
+			var expectedDemand = expectedDemandFunction.applyAsInt(zone);
+			if (expectedDemand == 0) {
+				return 0;
+			}
+			return (int)Math.round(params.getTargetAlpha() * expectedDemand + params.getTargetBeta());
+		};
 	}
 }
